api to get all times

+++ src/models/shape.go
@@ -2,8 +2,8 @@ package models
 
 type Shape struct{
 	ShapeId 			string 		`json:"shapeId"`
-	ShapePtLat			float32		`json:"shapePtLat"`
-	ShapePtLon			float32 	`json:"shapePtLon"`
+	ShapePtLat			float64		`json:"shapePtLat"`
+	ShapePtLon			float64 	`json:"shapePtLon"`
 	ShapePtSequence 	int 		`json:"shapePtSequence"`
 	ShapeDistTraveled 	int 		`json:"shapeDistTraveled"`
 }
